+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## unreleased
+### Added
+* Added an option to merge wimdo and camdo armatures on import. Leave this unchecked for editing and exporting models.
+
 ### Changed
 * Changed imported material names to use the model name as a prefix to avoid name conflicts across imports.